# api/docs/openapi.yaml
openapi: 3.0.1
info:
  contact: {}
  description: "API for managing sandboxes, organizations, billing, and hosts"
  title: Fluid API
  version: "1.0"
servers:
- url: //localhost:8081/v1
paths:
  /auth/github:
    get:
      description: Redirect to GitHub OAuth authorization page
      responses:
        "302":
          content: {}
          description: Redirect to GitHub
        "501":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/rest.swaggerError"
          description: Not Implemented
      summary: GitHub OAuth login
      tags:
      - Auth
  /auth/github/callback:
    get:
      description: "Handle GitHub OAuth callback, create or link user, set session\
        \ cookie, and redirect to dashboard"
      parameters:
      - description: OAuth authorization code
        in: query
        name: code
        required: true
        schema:
          type: string
      responses:
        "302":
          content: {}
          description: Redirect to dashboard
        "400":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/rest.swaggerError"
          description: Bad Request
        "500":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/rest.swaggerError"
          description: Internal Server Error
      summary: GitHub OAuth callback
      tags:
      - Auth
  /auth/google:
    get:
      description: Redirect to Google OAuth authorization page
      responses:
        "302":
          content: {}
          description: Redirect to Google
        "501":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/rest.swaggerError"
          description: Not Implemented
      summary: Google OAuth login
      tags:
      - Auth
  /auth/google/callback:
    get:
      description: "Handle Google OAuth callback, create or link user, set session\
        \ cookie, and redirect to dashboard"
      parameters:
      - description: OAuth authorization code
        in: query
        name: code
        required: true
        schema:
          type: string
      responses:
        "302":
          content: {}
          description: Redirect to dashboard
        "400":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/rest.swaggerError"
          description: Bad Request
        "500":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/rest.swaggerError"
          description: Internal Server Error
      summary: Google OAuth callback
      tags:
      - Auth
  /auth/login:
    post:
      description: "Authenticate with email and password, returns a session cookie"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/rest.loginRequest"
        description: Login credentials
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/rest.authResponse"
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/rest.swaggerError"
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/rest.swaggerError"
          description: Unauthorized
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/rest.swaggerError"
          description: Internal Server Error
      summary: Log in
      tags:
      - Auth
      x-codegen-request-body-name: request
  /auth/logout:
    post:
      description: Invalidate the current session and clear the session cookie
      responses:
        "200":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: OK
      security:
      - CookieAuth: []
      summary: Log out
      tags:
      - Auth
  /auth/me:
    get:
      description: Return the currently authenticated user
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/rest.authResponse"
          description: OK
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/rest.swaggerError"
          description: Unauthorized
      security:
      - CookieAuth: []
      summary: Get current user
      tags:
      - Auth
  /auth/register:
    post:
      description: Create a new user account and return a session cookie
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/rest.registerRequest"
        description: Registration details
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/rest.authResponse"
          description: Created
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/rest.swaggerError"
          description: Bad Request
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/rest.swaggerError"
          description: Conflict
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/rest.swaggerError"
          description: Internal Server Error
      summary: Register a new user
      tags:
      - Auth
      x-codegen-request-body-name: request
  /billing/calculator:
    post:
      description: Calculate estimated monthly costs based on resource usage
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/rest.calculatorRequest"
        description: Resource quantities
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/rest.calculatorResponse"
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/rest.swaggerError"
          description: Bad Request
      summary: Pricing calculator
      tags:
      - Billing
      x-codegen-request-body-name: request
...
  /orgs:
    get:
      description: List all organizations the current user belongs to
      responses:
        "200":
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
          description: OK
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/rest.swaggerError"
          description: Internal Server Error
      security:
      - CookieAuth: []
      summary: List organizations
      tags:
      - Organizations
    post:
      description: Create a new organization and add the current user as owner
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/rest.createOrgRequest"
        description: Organization details
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/rest.orgResponse"
          description: Created
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/rest.swaggerError"
          description: Bad Request
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/rest.swaggerError"
          description: Conflict
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/rest.swaggerError"
          description: Internal Server Error
      security:
      - CookieAuth: []
      summary: Create organization
      tags:
      - Organizations
      x-codegen-request-body-name: request
  /orgs/{slug}:
    delete:
      description: Delete an organization (owner only)
      parameters:
      - description: Organization slug
        in: path
        name: slug
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: OK
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/rest.swaggerError"
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/rest.swaggerError"
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/rest.swaggerError"
          description: Internal Server Error
      security:
      - CookieAuth: []
      summary: Delete organization
      tags:
      - Organizations
    get:
      description: Get organization details by slug
      parameters:
      - description: Organization slug
        in: path
        name: slug
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/rest.orgResponse"
          description: OK
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/rest.swaggerError"
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/rest.swaggerError"
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/rest.swaggerError"
          description: Internal Server Error
      security:
      - CookieAuth: []
      summary: Get organization
      tags:
      - Organizations
    patch:
      description: Update organization details (owner or admin only)
      parameters:
      - description: Organization slug
        in: path
        name: slug
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/rest.updateOrgRequest"
        description: Fields to update
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/rest.orgResponse"
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/rest.swaggerError"
          description: Bad Request
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/rest.swaggerError"
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/rest.swaggerError"
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/rest.swaggerError"
          description: Internal Server Error
      security:
      - CookieAuth: []
      summary: Update organization
      tags:
      - Organizations
      x-codegen-request-body-name: request
...
components:
  schemas:
...
    rest.authResponse:
      example:
        user:
          email_verified: true
          avatar_url: avatar_url
          id: id
          display_name: display_name
          email: email
      properties:
        user:
          $ref: "#/components/schemas/rest.userResponse"
      type: object
...
    rest.calculatorRequest:
      properties:
        agent_hosts:
          type: integer
        concurrent_sandboxes:
          type: integer
        hours_per_month:
          type: number
        source_vms:
          type: integer
      type: object
    rest.calculatorResponse:
      example:
        sandbox_cost: 6.027456183070403
        agent_host_cost: 0.8008281904610115
        source_vm_cost: 1.4658129805029452
        currency: currency
        total_monthly: 5.962133916683182
      properties:
        agent_host_cost:
          type: number
        currency:
          type: string
        sandbox_cost:
          type: number
        source_vm_cost:
          type: number
        total_monthly:
          type: number
      type: object
...
    rest.createOrgRequest:
      properties:
        name:
          type: string
        slug:
          type: string
      type: object
...
    rest.loginRequest:
      properties:
        email:
          type: string
        password:
          type: string
      type: object
...
    rest.orgResponse:
      example:
        stripe_customer_id: stripe_customer_id
        owner_id: owner_id
        name: name
        created_at: created_at
        id: id
        slug: slug
      properties:
        created_at:
          type: string
        id:
          type: string
        name:
          type: string
        owner_id:
          type: string
        slug:
          type: string
        stripe_customer_id:
          type: string
      type: object
    rest.registerRequest:
      properties:
        display_name:
          type: string
        email:
          type: string
        password:
          type: string
      type: object
    rest.swaggerError:
      example:
        code: 0
        details: details
        error: error
      properties:
        code:
          type: integer
        details:
          type: string
        error:
          type: string
      type: object
    rest.updateOrgRequest:
      properties:
        name:
          type: string
      type: object
...
    rest.userResponse:
      example:
        email_verified: true
        avatar_url: avatar_url
        id: id
        display_name: display_name
        email: email
      properties:
        avatar_url:
          type: string
        display_name:
          type: string
        email:
          type: string
        email_verified:
          type: boolean
        id:
          type: string
      type: object
...